Moved Tests to their own Class for better navigation Also, catch exception thrown in SetTempUrlKey if disk doesn't exists
