+++ src/Commands/SetTempUrlKey.php
@@ -5,6 +5,7 @@ namespace Sausin\LaravelOvh\Commands;
 use Exception;
 use Illuminate\Console\Command;
 use Illuminate\Support\Facades\Storage;
+use InvalidArgumentException;
 use OpenStack\ObjectStore\v1\Models\Container;
 
 class SetTempUrlKey extends Command
@@ -48,7 +49,7 @@ class SetTempUrlKey extends Command
     {
         try {
             $this->container = Storage::disk($this->option('disk'))->getAdapter()->getContainer();
-        } catch (\InvalidArgumentException $e) {
+        } catch (InvalidArgumentException $e) {
             $this->error($e->getMessage());
 
             return;
