+++ tests/Functional/UrlGenerationTest.php
@@ -1,0 +1,48 @@
+<?php
+
+namespace Sausin\LaravelOvh\Tests\Functional;
+
+use Carbon\Carbon;
+use Sausin\LaravelOvh\Tests\TestCase;
+
+class UrlGenerationTest extends TestCase
+{
+    public function testUrlMethod()
+    {
+        $this->object->shouldNotReceive('retrieve');
+        $this->container->shouldNotReceive('getObject');
+
+        $url = $this->adapter->getUrl('hello');
+
+        $this->assertEquals('https://storage.TestingGround.cloud.ovh.net/v1/AUTH_AwesomeProject/my-container/hello', $url);
+    }
+
+    public function testUrlConfirmMethod()
+    {
+        $this->object->shouldReceive('retrieve')->once();
+        $this->object->name = 'hello/world';
+        $this->object->lastModified = date('Y-m-d');
+        $this->object->contentType = 'mimetype';
+        $this->object->contentLength = 1234;
+
+        $this->container
+            ->shouldReceive('getObject')
+            ->once()
+            ->with('hello')
+            ->andReturn($this->object);
+
+        $url = $this->adapter->getUrlConfirm('hello');
+
+        $this->assertEquals('https://storage.TestingGround.cloud.ovh.net/v1/AUTH_AwesomeProject/my-container/hello', $url);
+    }
+
+    public function testTemporaryUrlMethod()
+    {
+        $this->object->shouldNotReceive('retrieve');
+        $this->container->shouldNotReceive('getObject');
+
+        $url = $this->adapter->getTemporaryUrl('hello.jpg', Carbon::now()->addMinutes(10));
+
+        $this->assertNotNull($url);
+    }
+}
